vaihe1

+++ src/api/models/tagModel.ts
@@ -1,16 +1,13 @@
 import {ResultSetHeader, RowDataPacket} from 'mysql2';
-import {TagResult} from '@sharedTypes/DBTypes';
+import {Tag, TagResult} from '@sharedTypes/DBTypes';
 import promisePool from '../../lib/db';
-import {fetchData} from '../../lib/functions';
 import {MessageResponse} from '@sharedTypes/MessageTypes';
 
 // Request a list of tags
-const fetchAllTags = async (): Promise<TagResult[] | null> => {
+const fetchAllTags = async (): Promise<Tag[] | null> => {
   try {
-    const [rows] = await promisePool.execute<RowDataPacket[] & TagResult[]>(
-      `SELECT Tags.tag_id, Tags.tag_name, MediaItemTags.media_id
-       FROM Tags
-       JOIN MediaItemTags ON Tags.tag_id = MediaItemTags.tag_id`
+    const [rows] = await promisePool.execute<RowDataPacket[] & Tag[]>(
+      'SELECT * FROM Tags'
     );
     if (rows.length === 0) {
       return null;
@@ -22,58 +19,46 @@ const fetchAllTags = async (): Promise<TagResult[] | null> => {
   }
 };
 
-// Post a new tag
-const postTag = async (
-  tag: Omit<TagResult, 'tag_id'>
-): Promise<MessageResponse | null> => {
-  const connection = await promisePool.getConnection();
+const fetchTagByName = async (tag_name: string): Promise<Tag | null> => {
   try {
-    await connection.beginTransaction();
-    const [tagResult] = await connection.execute<ResultSetHeader>(
+    const [rows] = await promisePool.execute<RowDataPacket[] & Tag[]>(
+      'SELECT * FROM Tags WHERE tag_name = ?',
+      [tag_name]
+    );
+    if (rows.length === 0) {
+      return null;
+    }
+    return rows[0];
+  } catch (e) {
+    console.error('fetchTagByName error', (e as Error).message);
+    throw new Error((e as Error).message);
+  }
+};
+
+// Post a new tag
+const postTag = async (tag_name: string): Promise<Tag | null> => {
+  try {
+    // check if tag already exists
+    const oldTag = await fetchTagByName(tag_name);
+    if (oldTag) {
+      return null;
+    }
+
+    const [tagResult] = await promisePool.execute<ResultSetHeader>(
       'INSERT INTO Tags (tag_name) VALUES (?)',
-      [tag.tag_name]
+      [tag_name]
     );
     if (tagResult.affectedRows === 0) {
       return null;
     }
 
-    const [mediaItemTagResult] = await connection.execute<ResultSetHeader>(
-      'INSERT INTO MediaItemTags (media_id, tag_id) VALUES (?, ?)',
-      [tag.media_id, tagResult.insertId]
+    const [rows] = await promisePool.execute<RowDataPacket[] & Tag[]>(
+      'SELECT * FROM Tags WHERE tag_id = ?',
+      [tagResult.insertId]
     );
-
-    await connection.commit();
-
-    if (mediaItemTagResult.affectedRows === 0) {
-      return null;
-    }
-
-    return {message: 'Tag created'};
+    return rows[0];
   } catch (e) {
-    await connection.rollback();
     console.error('postTag error', (e as Error).message);
-    throw new Error((e as Error).message);
-  } finally {
-    connection.release();
-  }
-};
-
-// Request a list of media items by tag
-const fetchMediaByTag = async (tag: string): Promise<TagResult[] | null> => {
-  try {
-    const [rows] = await promisePool.execute<RowDataPacket[] & TagResult[]>(
-      `SELECT Tags.tag_id, Tags.tag_name, MediaItemTags.media_id
-       FROM Tags
-       JOIN MediaItemTags ON Tags.tag_id = MediaItemTags.tag_id
-       WHERE Tags.tag_name = ?`,
-      [tag]
-    );
-    if (rows.length === 0) {
-      return null;
-    }
-    return rows;
-  } catch (e) {
-    console.error('fetchMediaByTag error', (e as Error).message);
     throw new Error((e as Error).message);
   }
 };
@@ -103,18 +88,24 @@ const deleteTag = async (id: number): Promise<MessageResponse | null> => {
   const connection = await promisePool.getConnection();
   try {
     await connection.beginTransaction();
-    const [tagResult] = await connection.execute<ResultSetHeader>(
-      'DELETE FROM Tags WHERE tag_id = ?',
-      [id]
-    );
-    if (tagResult.affectedRows === 0) {
-      return null;
-    }
 
     const [mediaItemTagResult] = await connection.execute<ResultSetHeader>(
       'DELETE FROM MediaItemTags WHERE tag_id = ?',
       [id]
     );
+
+    if (mediaItemTagResult.affectedRows === 0) {
+      return null;
+    }
+
+    const [tagResult] = await connection.execute<ResultSetHeader>(
+      'DELETE FROM Tags WHERE tag_id = ?',
+      [id]
+    );
+
+    if (tagResult.affectedRows === 0) {
+      return null;
+    }
 
     await connection.commit();
 
@@ -132,4 +123,4 @@ const deleteTag = async (id: number): Promise<MessageResponse | null> => {
   }
 };
 
-export {fetchAllTags, postTag, fetchMediaByTag, fetchTagsByMediaId, deleteTag};
+export {fetchAllTags, postTag, fetchTagsByMediaId, fetchTagByName, deleteTag};
